refactor(sound): restructure audio elements and enhance error handling in PlaySoundService
- Introduced a new module for audio elements, consolidating sound assets and improving organization.
- Updated PlaySoundService to utilize structured error handling with PlaySoundServiceError for better clarity and type safety.
- Adjusted return types in PlaySoundService to use Result, ensuring consistency across desktop and web implementations.
