+++ apps/app/src/lib/services/sound/PlaySoundService.desktop.ts
@@ -1,12 +1,23 @@
-import { Ok } from '@epicenterhq/result';
-import type { PlaySoundService } from './PlaySoundService';
-import { audioElements } from './audioElements';
+import { tryAsync } from '@epicenterhq/result';
+import type {
+	PlaySoundService,
+	PlaySoundServiceError,
+} from './PlaySoundService';
+import { audioElements } from './_audioElements';
 
 export function createPlaySoundServiceDesktop(): PlaySoundService {
 	return {
-		playSound: async (soundName) => {
-			await audioElements[soundName].play();
-			return Ok(undefined);
-		},
+		playSound: async (soundName) =>
+			tryAsync({
+				try: async () => {
+					await audioElements[soundName].play();
+				},
+				mapErr: (error): PlaySoundServiceError => ({
+					name: 'PlaySoundServiceError',
+					message: 'Failed to play sound',
+					context: { soundName },
+					cause: error,
+				}),
+			}),
 	};
 }
