+++ apps/app/src/lib/services/sound/PlaySoundService.ts
@@ -1,7 +1,10 @@
-import type { WhisperingResult, WhisperingSoundNames } from '@repo/shared';
+import type { Result, TaggedError } from '@epicenterhq/result';
+import type { WhisperingSoundNames } from '@repo/shared';
+
+export type PlaySoundServiceError = TaggedError<'PlaySoundServiceError'>;
 
 export type PlaySoundService = {
 	playSound: (
 		soundName: WhisperingSoundNames,
-	) => Promise<WhisperingResult<void>>;
+	) => Promise<Result<void, PlaySoundServiceError>>;
 };
